Import libshl with crit-trie implementation
libshl is a collection of small self-hosting C helpers. It is not intended
as library to be installed into systems. Instead, it is rather a place to
collect handy helpers which then can be easily copied into your
applications.

This repository contains extensive tests for all modules so we can try to
guarantee their stability and correctness.

The whole library is dedicated to the Public Domain. Copy it if you want.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>