trie: switch to caller-allocated entries
Instead of allocating entries in the trie itself, we now let the caller do
that. The syntax changes a bit as we now need a double-pointer to the
keys, but the semantics stay the same (you can see that from the tests).

Note that this still allows the caller to save any kind of data with the
keys. However, they're now responsible for allocating the key-ptr+data
objects.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>