Add githead helper
shl_githead is a short helper to put the build-git-revision into compiled
objects as string. It is bundled to a Makefile.am snippet that takes care
of it.

Note that it is highly optimized to only rebuild if the revision really
changed. Furthermore, it is created to work with dist-tarballs and VPATH
builds.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>