macro: add safe-multiplication helpers
Checking for overflows on additions is simple, however, for
multiplications it's non-trivial. In case we have a bigger integer type,
we can simply run the multiplication one class higher and check for an
overflow. However, for "unsigned long long" / "uint64_t" we need to use
a division to test for overflows beforehand.

It would be nice if the C language provided us with overflow register
flags. Unfortunately, there's no such thing so we're left with expensive
checks. Note that the division is horribly slow, so you should always
reconsider whether you really need this big space or whether other
heuristics could avoid most of the checks.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>